feat(orders): add Orders and Coupons with Minimal API + CQRS

+++ Models/ProductModel.cs
@@ -3,10 +3,10 @@ using System.ComponentModel.DataAnnotations;
 
 public class ProductModel
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Color { get; set; } = string.Empty;
 
-    public long ProductId { get; set; }
+    public int ProductId { get; set; }
     public Product? Product { get; set; }
 }
